pyproject.toml
[tool.poetry]
name = "se-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
langchain = "^0.1.20"
langchain-nvidia-ai-endpoints = "^0.0.14"
chromadb = "^0.5.0"
quart = "^0.19.5"
python-dotenv = "^1.0.1"
quart-cors = "^0.7.0"
nemoguardrails = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
